~Moved execution compilation to new file
